Refactor ProjectInfo into seperate class
